05 Lightroom AI Description Generation/requirements.txt
accelerate==0.26.1
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
Flask==3.0.1
fonttools==4.47.2
fsspec==2023.12.2
gradio==4.14.0
gradio_client==0.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jproperties==2.1.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
keyboard==0.13.5
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
orjson==3.9.12
packaging==23.2
pandas==2.1.4
paramiko==3.4.0
pillow==10.2.0
pillow-heif==0.14.0
psutil==5.9.7
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pysftp==0.2.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
rawpy==0.19.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.1
scipy==1.11.4
semantic-version==2.10.0
sftputil==1.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
sympy==1.12
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
uform==1.0.3
urllib3==2.1.0
uvicorn==0.26.0
websockets==11.0.3
Werkzeug==3.0.1
wikipedia==1.4.0
